+++ account-service/src/main/java/ru/test/best/dto/AccountDto.java
@@ -13,8 +13,8 @@ import java.io.Serializable;
 @Setter
 @RequiredArgsConstructor
 public class AccountDto implements Serializable {
-    private int clientId;
-    private int points;
+    private Long clientId;
+    private Long points;
     private String name;
     private String role;
 }
